# openapi.yml
openapi: 3.0.3
info:
  title: ACMS - Apartment Committee Management System
  description: |
    Multi-tenant SaaS platform for apartment/housing societies.
    Five role types: site_admin, apartment_admin, committee_head, committee_member, resident.
    Tenant isolation via `apartmentId` — all non-site-admin queries are scoped.
  version: 1.0.0
  contact:
    name: ACMS Team
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CommitteeRole:
      type: string
      enum: [committee_head, committee_member]
...
    LedgerType:
      type: string
      enum: [income, expense]
    BillStatus:
      type: string
      enum: [paid, unpaid, overdue]
...
    ComplaintStatus:
      type: string
      enum: [open, in_progress, resolved]

    # ────────── Auth ──────────
...
    Committee:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        name: { type: string }
        headUserId: { type: string, nullable: true, description: 'User ID or populated' }
        createdAt: { type: string, format: date-time }
        status: { type: string, enum: [active, inactive] }
    CreateCommitteeRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, example: Maintenance Committee }
...
    CommitteeMember:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        committeeId: { type: string }
        type: { $ref: '#/components/schemas/CommitteeRole' }
        name: { type: string }
        identifier: { type: string }
    CreateCommitteeMemberRequest:
      type: object
      required: [name, identifier, password]
      properties:
        name: { type: string }
        identifier: { type: string }
        password: { type: string, minLength: 6 }

    # ────────── Committee Ledger ──────────
    LedgerEntry:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        committeeId: { type: string }
        type: { $ref: '#/components/schemas/LedgerType' }
        amount: { type: number }
        description: { type: string }
        date: { type: string, format: date-time }
        recordedBy: { type: string, description: 'User ID or populated' }
    CreateLedgerEntryRequest:
      type: object
      required: [type, amount, description]
      properties:
        type: { $ref: '#/components/schemas/LedgerType' }
        amount: { type: number, exclusiveMinimum: 0 }
        description: { type: string }
        date: { type: string, format: date-time, description: Defaults to now }

    # ────────── Maintenance Bills ──────────
    MaintenanceBill:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        committeeId: { type: string, description: 'Committee ID or populated' }
        unitId: { type: string, description: 'Unit ID or populated' }
        amount: { type: number }
        period: { type: string }
        status: { $ref: '#/components/schemas/BillStatus' }
        dueDate: { type: string, format: date-time }
    GenerateBillsRequest:
      type: object
      required: [amount, period, dueDate]
      properties:
        amount: { type: number, exclusiveMinimum: 0 }
        period: { type: string, example: '2026-07' }
        dueDate: { type: string, format: date-time }

    # ────────── Complaints ──────────
    Complaint:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        committeeId: { type: string, description: 'Committee ID or populated' }
        raisedByUnitId: { type: string, description: 'Unit ID or populated' }
        description: { type: string }
        status: { $ref: '#/components/schemas/ComplaintStatus' }
        assignedTo: { type: string, nullable: true, description: 'User ID or populated' }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
        rating: { type: integer, minimum: 1, maximum: 5, nullable: true }
    CreateComplaintRequest:
      type: object
      required: [committeeId, description]
      properties:
        committeeId: { type: string }
        description: { type: string }
    UpdateComplaintRequest:
      type: object
      properties:
        status: { $ref: '#/components/schemas/ComplaintStatus' }
        assignedTo: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
    RateComplaintRequest:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }

    # ────────── Visitor Log ──────────
...
paths:
...
  /apartment/committees:
    get:
      tags: [Apartment Admin]
      summary: List committees in the apartment
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Committee'
    post:
      tags: [Apartment Admin]
      summary: Create a committee
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommitteeRequest'
      responses:
        '201': { description: Committee created }
        '409': { description: Committee name already exists }
  /apartment/committees/{id}:
    put:
      tags: [Apartment Admin]
      summary: Update committee name/status
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                status: { type: string, enum: [active, inactive] }
      responses:
        '200': { description: Committee updated }
        '404': { description: Committee not found }
    delete:
      tags: [Apartment Admin]
      summary: Delete a committee
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Committee deleted }
...
  /committees:
    get:
      tags: [Committee]
      summary: List committees (apartment_admin sees all, committee sees own)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Committee'
  /committees/{id}:
    get:
      tags: [Committee]
      summary: Get a single committee by ID
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Committee details }
        '404': { description: Committee not found }
  /committees/{id}/members:
    get:
      tags: [Committee]
      summary: List members of a committee (head + members)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommitteeMember'
    post:
      tags: [Committee]
      summary: Add a committee member
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommitteeMemberRequest'
      responses:
        '201': { description: Member added }
        '409': { description: Identifier already exists }
  /committees/{committeeId}/members/{memberId}:
    delete:
      tags: [Committee]
      summary: Remove a committee member
      security: [{ BearerAuth: [] }]
      parameters:
        - name: committeeId
          in: path
          required: true
          schema: { type: string }
        - name: memberId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Member removed }

  # ────────── Committee Ledger ──────────
  /committees/{id}/ledger:
    get:
      tags: [Committee]
      summary: List committee ledger entries
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerEntry'
    post:
      tags: [Committee]
      summary: Add a ledger entry (committee_head only)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLedgerEntryRequest'
      responses:
        '201': { description: Entry added }

  # ────────── Committee Bills ──────────
  /committees/{id}/bills:
    get:
      tags: [Committee]
      summary: List maintenance bills for a committee
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceBill'
  /committees/{id}/bills/generate:
    post:
      tags: [Committee]
      summary: Generate bills for all occupied units in a committee
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateBillsRequest'
      responses:
        '201': { description: Bills generated with count }
  /committees/bills/{billId}/pay:
    put:
      tags: [Committee]
      summary: Pay a bill (resident only, scoped to their unit)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: billId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Bill marked as paid }
        '404': { description: Bill not found or already paid }

  # ══════════════════════════════════════════════════
  # COMPLAINTS
  # ══════════════════════════════════════════════════
  /complaints:
    get:
      tags: [Complaints]
      summary: List complaints (scoped by role)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Complaint'
    post:
      tags: [Complaints]
      summary: Raise a complaint (resident only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComplaintRequest'
      responses:
        '201': { description: Complaint created }
        '400': { description: No unit associated }
        '403': { description: Only residents can raise complaints }
  /complaints/{id}:
    put:
      tags: [Complaints]
      summary: Update complaint status (committee head/member) or set rating
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateComplaintRequest'
      responses:
        '200': { description: Complaint updated }
        '404': { description: Complaint not found }
  /complaints/{id}/rate:
    put:
      tags: [Complaints]
      summary: Rate a resolved complaint (resident who raised it only)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateComplaintRequest'
      responses:
        '200': { description: Rating submitted }
        '403': { description: Not the complaint owner }
        '404': { description: Complaint not found or not resolved }

  # ══════════════════════════════════════════════════
  # VISITOR LOGS
  # ══════════════════════════════════════════════════
...